+++ basic/controllers/FlightController.php
@@ -78,8 +78,13 @@ class FlightController extends Controller
      */
     public function actionView($id)
     {
+        $model = $this->findModel($id);
+        $pob = ($model->pax_adults !== null)
+            ? $model->pax_adults + $model->pax_children + $model->aircraft->aircraftConfiguration->crew
+            : 'X';
         return $this->render('view', [
-            'model' => $this->findModel($id),
+            'model' => $model,
+            'pob'   => $pob,
         ]);
     }