+++ basic/views/submitted-flight-plan/update.php
@@ -21,6 +21,7 @@ $this->params['breadcrumbs'][] = Yii::t('app', 'Update');
         'entity' => $entity,
         'pilotName' => $model->pilot->fullname,
         'mode' => 'update',
+        'pob' => $pob,
     ]) ?>
 
 </div>